# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: PawPals API
  version: 1.0.0
  description: |
    REST API for the PawPals mobile app. The frontend must not access the database
    or Supabase directly; all data access goes through this API layer.
    Provides CRUD operations for profiles, dogs, matches, playdates, ratings, and messages.
servers:
  - url: https://api.pawpals.example.com/v1
    description: Production server (example)
  - url: http://localhost:4000/v1
    description: Local dev server
...
paths:
  /profiles/me:
    get:
      tags: [Profiles]
      summary: Get current user's profile
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Profile not found
  /profiles:
    post:
      tags: [Profiles]
      summary: Create or upsert current user's profile
      description: |
        Creates a profile for the authenticated user. The `id` is taken from
        the authenticated user (auth.users.id); the client does not supply it.
      operationId: createProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreate'
      responses:
        '201':
          description: Created profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /profiles/{userId}:
    get:
      tags: [Profiles]
      summary: Get profile by user ID
      operationId: getProfileById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Profile not found
    patch:
      tags: [Profiles]
      summary: Update profile
      operationId: updateProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Profile not found
    delete:
      tags: [Profiles]
      summary: Delete profile
      operationId: deleteProfile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: Profile deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /ratings:
    get:
      tags: [Ratings]
      summary: List ratings for current user
      description: Returns ratings authored by the current user.
      operationId: listRatings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Ratings]
      summary: Create a rating for a playdate
      operationId: createRating
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingCreate'
      responses:
        '201':
          description: Created rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /ratings/{ratingId}:
    get:
      tags: [Ratings]
      summary: Get a rating by ID
      operationId: getRatingById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RatingIdParam'
      responses:
        '200':
          description: Rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Rating not found
    patch:
      tags: [Ratings]
      summary: Update a rating
      operationId: updateRating
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RatingIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingUpdate'
      responses:
        '200':
          description: Updated rating
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Rating not found
    delete:
      tags: [Ratings]
      summary: Delete a rating
      operationId: deleteRating
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RatingIdParam'
      responses:
        '204':
          description: Rating deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserIdParam:
      in: path
      name: userId
      required: true
      schema:
        type: string
        format: uuid
      description: ID from auth.users / profiles.id
...
    RatingIdParam:
      in: path
      name: ratingId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized – missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden – user not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          nullable: true
        email_verified:
          type: boolean
          nullable: true
        join_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, created_at, updated_at]
    ProfileCreate:
      type: object
      properties:
        full_name:
          type: string
        email_verified:
          type: boolean
      required: [full_name]
    ProfileUpdate:
      type: object
      properties:
        full_name:
          type: string
        email_verified:
          type: boolean
...
    Rating:
      type: object
      properties:
        rating_id:
          type: string
          format: uuid
        playdate_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - rating_id
        - playdate_id
        - user_id
        - stars
        - created_at
        - updated_at
    RatingCreate:
      type: object
      properties:
        playdate_id:
          type: string
          format: uuid
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - playdate_id
        - stars
    RatingUpdate:
      type: object
      properties:
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          example: Something went wrong
        details:
          type: object
          additionalProperties: true